Move GetLanguageKind(...) implementation methods to a helper class
